requirements.txt
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.7
durationpy==0.9
google-auth==2.37.0
idna==3.10
kubernetes==31.0.0
oauthlib==3.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
six==1.17.0
urllib3==2.3.0
websocket-client==1.8.0
